pyproject.toml
[tool.poetry]
name = "nowdjango"
version = "0.1.0"
description = ""
authors = ["AlexGanechkin <terracat80@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
